refactor(shared): マイグレーションコードをドメインスキーマから分離
index.ts にドメインスキーマとマイグレーションコードが混在していたため,
3ファイル構成に分割して関心を分離する.

- schemas.ts: ドメインスキーマのみ (NodeLayout, GraphNode, Sheet 等)
- migrations.ts: バージョン別スキーマ (V1/V2) とマイグレーション関数
- index.ts: 両者の re-export + HTTP API スキーマ

migrations.ts は schemas.ts から import するため循環依存なし.
将来バージョンを追加する際は migrations.ts のみを変更すればよい.

Confidence: high
Scope-risk: narrow
